requirements.txt
opencv_contrib_python==4.2.0.32
moment==0.8.2
getmac==0.8.2
bugsnag==3.6.0
numpy==1.17.2
tensorflow==2.2.0
paho_mqtt==1.5.0
Pillow==7.1.2
pyserial==3.4
pyzmq==19.0.1
